+++ CHANGELOG.md
@@ -23,7 +23,7 @@ omnibus-gitlab repository.
 - Add a command to promote secondary node to primary for GitLab Geo
 - Fixed behaviour of postgres_user provider with unspecified passwords/options 8bd2e615
 - Added roles to ease HA Postgres configuration d7d0f32a
-
+- Add a runit service to probe repository storages
 
 10.2.3